cchore(): move styles to file
