+++ src/components/CheckIn/CheckIn.js
@@ -1,13 +1,12 @@
 import React, { Component } from 'react'
 import {
-  Platform,
-  StyleSheet,
   Text,
   View,
   Button
 } from 'react-native'
+import styles from './styles/styles'
 
-export default class CheckIn extends Component {
+class CheckIn extends Component {
   render() {
     return (
       <View style={styles.container}>
@@ -25,21 +24,4 @@ export default class CheckIn extends Component {
   }
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#F5FCFF',
-  },
-  welcome: {
-    fontSize: 20,
-    textAlign: 'center',
-    margin: 10,
-  },
-  instructions: {
-    textAlign: 'center',
-    color: '#333333',
-    marginBottom: 5,
-  },
-})
+export default CheckIn
