+++ src/components/NewCheckin/NewCheckin.js
@@ -1,13 +1,12 @@
 import React, { Component } from 'react'
 import {
-  Platform,
-  StyleSheet,
   Text,
   View,
   Button
 } from 'react-native'
+import styles from './styles/styles'
 
-export default class NewCheckIn extends Component {
+class NewCheckIn extends Component {
   render() {
     return (
       <View style={styles.container}>
@@ -31,21 +30,4 @@ export default class NewCheckIn extends Component {
   }
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#F5FCFF',
-  },
-  welcome: {
-    fontSize: 20,
-    textAlign: 'center',
-    margin: 10,
-  },
-  instructions: {
-    textAlign: 'center',
-    color: '#333333',
-    marginBottom: 5,
-  },
-})
+export default NewCheckIn
